requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.5
anyio==3.5.0
APScheduler==3.8.1
asgiref==3.4.1
astroid==2.9.3
asttokens==2.0.5
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
bcrypt==3.2.0
black==21.12b0
ccxt==1.69.18
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cryptography==36.0.1
cycler==0.11.0
databases==0.5.5
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.3
executing==0.8.2
fastapi==0.72.0
FastAPI-SQLAlchemy==0.2.1
fonttools==4.28.5
frozenlist==1.3.0
graphviz==0.19.1
greenlet==1.1.2
h11==0.13.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.13.2
hyperframe==6.0.1
idna==3.3
importlib-metadata==4.10.1
importlib-resources==5.4.0
ipykernel==6.7.0
ipython==8.0.1
isort==5.10.1
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.1.1
jupyter-core==4.9.1
kiwisolver==1.3.2
lazy-object-proxy==1.7.1
Mako==1.1.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
multidict==5.2.0
mypy-extensions==0.4.3
nest-asyncio==1.5.4
nltk==3.6.7
numpy==1.22.1
oauthlib==3.1.1
packaging==21.3
pandas==1.3.5
parso==0.8.3
passlib==1.7.4
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.4.1
priority==2.0.0
prompt-toolkit==3.0.24
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.1
pycares==4.1.2
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
Pygments==2.11.2
PyJWT==2.3.0
pylint==2.12.2
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.3.0
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
stack-data==0.1.4
starlette==0.17.1
textblob==0.17.1
threadpoolctl==3.0.0
toml==0.10.2
tomli==1.2.3
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
tweepy==4.4.0
typing-extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
uvicorn==0.17.0
wcwidth==0.2.5
wrapt==1.13.3
wsproto==1.0.0
yarl==1.7.2
zipp==3.7.0
